Makes me soundtainer (#89405)
adds me to the sound folder codeowners
I've done a lot of changes to the sound folder and I'd like to be
notified when people open a pr changing sound stuff, so I can review or
give tips.
In fact, 15% of all merged prs with the sound label in the codebase were
made by me.

![image](https://github.com/user-attachments/assets/f3ac8b71-6052-4377-ae35-e2de809a9daa)

![image](https://github.com/user-attachments/assets/0ddcf525-d185-4c20-9e92-cfd7fdd1d70b)