Lost heretic blades can be tracked with the living heart (#89544)
## About The Pull Request
Lost heretic blades can be tracked with the living heart
## Why It's Good For The Game
While this is meant to be punishing, having your blades hidden from you
is just not a fun experience at all for the heretic and is kinda lame.

Making them trackable is not a get-out-of-consequences free card, but it
does mean that you at least get a fighting chance at recovering a
portion of your power.