[NO GBP] Fixing trophy fishes not loading mats (#89623)
## About The Pull Request
For some reason, that either wasn't an issue before, or it always was
and I did not fix it. Either way, `List()` is a proc and `material_path`
variable is being read as an arg called "material_path" which is what
the key is being set to, instead of the stored value of the variable, I
think.

## Why It's Good For The Game
Fixing stuff.

## Changelog

:cl:
fix: Fixing loaded trophy fishes possibly not having materials.
/:cl: