[NO-GBP] Stops the charging defib mount from runtiming when there's no defib (#89688)
## About The Pull Request

Let this be a lesson to check your runtimes when testing your own
bugfixes.

I simply assumed that the charging defib mount only processes with a
defib in it (This is how the process() was written before #89299
anyhow). Seems this is not the case for some reason.
## Why It's Good For The Game
1 less runtime

I promise Ill come back to refactor this hellish process() timewaster.
## Changelog
:cl:
fix: The charging defib mount no longer runtimes when there's no defib
/:cl: