Contractor Kit out of lowpop  (#89757)
## About The Pull Request

Contractor Kit is only available on a population greater than the config
defined lowpop number.

## Why It's Good For The Game

Antags on lowpop are impossible to balance, since there's usually few or
no security or headstaff. Contractor Kit allows access to extremely
strong tools that give a huge advantage in one-on-one combat, meaning
that it's very easy and common for them to basically run unopposed in
lowpop. This is happening more now, because of the removal of progtot
has made the contractor kit a more popular option, being the remaining
progression based choice for traitors.

Lowpop murderboning is against the rules now on all servers, for good
reason. And the reason why I think Contractor Kit doesn't belong on
lowpop either is for very similar reasons. In higher pop where there is
adequate security, it's good enough to have antags that create conflict
and incentive to capture or kill them, and they don't really need to be
creative about how they antag. In lowpop though, the lack of meaningful
opponents for an antag makes it functionally an auto-win. The problem
with Contractor Kit in particular, is that for everyone else on the
station, it's a misery-generating machine. It makes Cargo impossible to
play as you will constantly have 0 credits, and everyone on the station
is regularly getting their 5 minute timeouts to the punishment zone.

The fun of antags for the station as a whole is in the struggle for who
will win, not an antag winning on minute 1 and subjecting everyone else
to your shit until they call the shuttle. Even an antag who has won can
theoretically be interesting with a gimmick, but much like how ProgTot
incentivized traitors to play in a boring manner, Contractor Kit
mechanically incentivizes people who take it into playing in a boring
way. Some people think that the kit is too powerful, but I think a good
start right now is simply to get it out of lowpop.

## Changelog

:cl:
balance: Contractor Kit is no longer available in lowpop.
/:cl: