Fix hauntinum reagent to work when exposed to item objects (#89734)
## About The Pull Request
Hauntium would cause a bunch of runtimes if it was exposed to objects on
a tile that weren't items. (structures, machinery, etc.) This was due to
the element rejecting non-item objects and forcing a crash message.

It also had used the wrong reagent volume settings to determine when it
should remove the effect. This caused it to instantly remove the haunted
effect as soon as it was applied.

Although I would have loved to adjust the length of time per reagent to
not be stupidly OP (20 secs per 1u) I feel like that's touching balance.

## Why It's Good For The Game
Code that breaks is bad.

## Changelog
:cl:
fix: Fix hauntinum reagent to work when it is exposed to item objects.
/:cl: