The rest of the lewd module I could stomach porting (#32)
There's a handful (quite a lot) of stuff missing but idc. this is as
good as it's getting on this timeframe. note the shoes have the same
bugs and issues other digi clothing has

:cl:
image: Nearly all of the lewd module is GAGs and rewritten now, and the
gasmask has been touched up a bit again.
/:cl: