Sapientsimov (#34)
## About The Pull Request
Implements an alternative lawset that preserves silicon's place as a
third party between most antagonists and the station : Sapientsimov.
## Why It's Good For The Game
TL;DR, most alternative lawsets on furry servers kinda just. Forget what
silicon's place in the game loop actually is and treat them as crew+,
or, worse, as security's puppets via crewsimov - which is really lame
for their ability to drive the round forward and create a bunch of
pettier conflicts that keep the game engaging. This addresses the
problem by instead going squarely in the other direction; having
protections previously afforded only to humanity now be afforded to all
sapient beings... Which, naturally, is also afforded to the same
antagonists that Asimov has always been; alongside a few extra.
## Changelog
:cl:
config: The default lawset has changed to sapientsimov via config.
/:cl: