Adds several new decals to the Tile Sprayer (#89883)
## About The Pull Request
Continuing my conquest against irrepairable features.
Today, adding a lot of the missing decals to the Tile Sprayer.
This doesn't include sidings or the likes, only the paints. ~~And it's
technically missing like, 3, but if they were added then the UI would
use the multiple-of-5 thats holding it together.~~

Also a tiny thing but I moved the 1-dir tiles to be the first of their
cluster, so it flows kind of like subsections (Primary, Filled Trims,
Thin Trims, so on so on)

> **Thank you @RealWinterFrost for showing me how easy it was to add
more decals to this list.**
> **Thank you @jlsnow301 for helping me with the UI when I decided that
I couldn't accept it being easy.**

<details><summary>A few preview tiles to prove that it works</summary>


![image](https://github.com/user-attachments/assets/086318f0-488b-4290-a7c4-7077774f20dd)

</details>
<details><summary>The UI now includes a scrollbar! Wow! This took me
hours!</summary>


![image](https://github.com/user-attachments/assets/10f769f9-51a5-4df6-9855-adbefdcd4357)

</details>
<details><summary>The full list of newly-available tile decals</summary>


![image](https://github.com/user-attachments/assets/0de41fe0-80a5-4288-8117-5f1d325cfa22)

</details>

> _The current way the spritesheets are set up, all tiles use the same
base floortile for the preview icon - this makes the Diagonal
Centre/Edge, Full Tile, Full Half, and Full Anticorner decals look a bit
odd in the UI. Technically functional, but- well, you can see it in the
3rd dropdown._
> _If anybody knows anything about spritesheets that might fix this then
I'd appreciate it._

> _~~Also if any really cool coders want to redo this UI with Categories
ala the RPD so we can cram the Tile Sprayer and Decal Painter into one
item that would be cool. I sadly cannot because I lack talent.~~_
## Why It's Good For The Game
One of the banes of SS13 is items or details players can use/see, but
cannot recreate. Giving players tools to rebuild what mappers place
makes the game world feel a lot more natural.

Now, our floors shall truly be beautiful.
## Changelog
:cl:
add: the Tile Sprayer now includes quite a few more tile decals! Warning
lines, guide arrows, diagonals and filled tiles! Wow!
/:cl: