Locational Blurbs (#52)
## About The Pull Request
Adds Locational Blurbs upon joining the game or moving from area to
area. This is a port and partial rewrite of
https://github.com/DaedalusDock/daedalusdock/pull/201 (and subsequently
https://github.com/NebulaSS13/Nebula/pull/357).

![image](https://github.com/user-attachments/assets/d4798f81-117d-4921-8335-fa3bc0908965)
## Why It's Good For The Game
Confers game information in a way that; while not diegetic, is
comfortable and unobtrusive.
## Changelog
:cl: Kapu1178 for Daedalus Dock, comma for Nebula, unit0016
add: Location Blurbs now greet you upon joining the game or entering a
new area.
/:cl: