Fixes some bitrunning mapping issues (#90411)
## About The Pull Request

Changed the turfs in the snow safehouse so it actually has atmosphere

Remapped the island brawl exit area to make the turfs line up with the
safehouse area, and better show where to drop lockboxes.

## Why It's Good For The Game

Bitrunners like being able to breathe, and map readability is good.

## Changelog
:cl:
fix: Bitrunners on Glacier Grind can now breathe in the safehouse.
fix: Bitrunners on Island Brawl will now be able to more clearly see the
exit area.
/:cl: