Adds the captain to the security department list. (#90504)
## About The Pull Request

Does what it says in the title.

## Why It's Good For The Game

In very many instances, captain gets slapped with security
responsibilities either during a security deficit, or as an authority in
security matters.

We actually use department lists to determine whether or not a role is
included in things such as our banning panel under certain grouped
roles, like security.

Rather than create a potential problem that can result from this, we can
just include captain in the security list so that they get grouped
together with security.

## Changelog
:cl:
admin: Adds Security as one of the captain's departments, ensuring that
departmental bans also include the captain, who is an effective member
of security.
/:cl: