Lobby info is now part of the hud instead of stat panel (#90572)
## About The Pull Request

Removes ``/mob/dead/get_status_tab_items()``, replaces it with a lobby
hud icon showing the same information

TV sprite made by Kryson, scanline and static taken from ``effects.dmi``
which I then blew up to fit the size.

https://github.com/user-attachments/assets/99733ae6-c596-42b1-bcae-d1a8d8f094c3

## Why It's Good For The Game

I want to try to remove as much reliance from the stat panel as possible
so we can hopefully remove it, as it takes up 1/3 of the screen for
nothing.

## Changelog

:cl: Kryson, JohnFulpWillard
qol: Lobby info is now on the title screen rather than being hidden in
the stat panel.
/:cl:
# Conflicts:
#	code/modules/mob/dead/dead.dm