Removes Blob HUD as a var on every single hud (#90588)
## About The Pull Request

Blob Overminds & Blobbernauts use a HUD that shows them the overmind's
health, which is currently a var called ``blobpwrdisplay`` and exists on
every single HUD in the game. This moves them to the respective mobs,
removing the need to do this.

## Why It's Good For The Game

Modularizes HUDs a bit more and removes a bad special var on base hud
datum. pmo.

## Changelog

Nothing player-facing.