Abductor In-Game Paper Guide Tweak (#90800)
## About The Pull Request

Encountered during a round in a downstream fork, the in-game abductor
guide is a tiny bit outdated (two steps are the wrong way around) and
also does not work on robot torsos (the kind you get from augmentation
or chaplain shenanigans).
It could be argued that this is an oversight and that abductors should
be able to operate on robotic torsos as normal, but that seems like a
different conversation. I've at least updated the guide to let abductors
know what to do with this unlikely scenario, as you can't get rid of
your test subjects unless you get a gland in them.

## Why It's Good For The Game

I think it's better when the in-game documentation is up to date.

## Changelog

:cl:
qol: Added some help for abductors if they get a weird half-machine test
subject in their ship.
/:cl: