RPD ui refactor (#91018)
## About The Pull Request
Fixes a possible bluescreen in the RPD. Issue was caused when the
backend didn't send the static data that the UI relied upon to populate
pipe layers (perhaps due to throttling)

Reworks the UI a bit so that the content section is scrollable instead
of the entire window. This prevents the top content from disappearing if
you're scrolling available pipes. Adds some extra typing where it was
type: any

![image](https://github.com/user-attachments/assets/d8e4e0b6-1798-4c76-bf94-10a50f916b3b)

## Why It's Good For The Game
Fixes #90992
## Changelog
:cl:
fix: Fixed an uncommon bluescreen in the rapid pipe dispenser UI.
/:cl:
# Conflicts:
#	tgui/packages/tgui/interfaces/RapidPipeDispenser.tsx