Fixes PDA notepad program always showing a blank field upon being opened. (#91157)
## About The Pull Request

Does what it says on the tin.
A previous pr seems to have accidentally set this to an empty string
when changing it, despite this needing to be set to `note` to work.
This fixes that.
## Why It's Good For The Game

less jank
## Changelog
:cl:
fix: PDA notepad app no longer always shows a blank field upon being
opened.
/:cl: