Fixes some issues with custom shuttles' initial docks (#91312)
## About The Pull Request

The initial dock for a custom shuttle - created when the shuttle is -
had the name of the shuttle area it was created in, rather than the name
of the area underneath the shuttle. Additionally, this dock was not
deleted when the shuttle left, preventing it from setting down a dock
back at that initial location. This PR fixes both of those issues.

## Why It's Good For The Game

Fixes an issue someone on Manuel couldn't be bothered to open on github
due to lack of motivation.

## Changelog

:cl:
fix: An initially created custom shuttle will no longer state it is
located at some place within itself.
fix: Custom shuttles will no longer have their docking computers blocked
by the area they were initially created at.
/:cl: