Fixes ninja camera vision module so it provides welding protection (#91428)
## About The Pull Request

The ninja camera vision module now provides the same welding protection
as the welding module, as intended.

This was due to the camera vision module not inheriting the function
which applied the welding protection in on_part_activation()

## Why It's Good For The Game

The camera vision module was not providing any flash protection to the
ninja, despite its description clearly stating;

"...as well as automatically dimming incoming flashes of light to
protect the user's eyes. Become the unseen."

In fact, the night vision goggles provided to the ninjas made them weak
to flashes.

## Changelog
:cl:
fix: Ninja Camera Vision module now correctly applies welding
protection.
/:cl: