fixes create_mob and create_turf, in line with create_object (#91518)
## About The Pull Request
As detailed in #91431, PR #91379 misplaced a few semicolons and broke
the functionality for these three admin tools. Only the object panel was
fixed, but the mob and turf panels remain broken. This fixes it, the
exact same issue was/is the problem.


## Why It's Good For The Game
Restores admin functionality that is currently missing.

## Changelog
:cl:
fix: create mob and create turf both function correctly again.
/:cl: