Allows attaching breath masks and tanks to operating tables (#91112)
## About The Pull Request

Air tanks (ones that can slot into tank holders) and breath masks
(normal, surgical and muzzle, which has been repathed to be a subtype of
breath masks instead of being a subtype of a muzzle, for some reason)
can now be attached to operating tables, and removed using
wrenches/screwdrivers respectively.
Right-clicking on an operating table with a mask attached will put it in
your hand akin to defib paddles (similarly snapping back if you walk
away). Dragging the table with an attached tank onto the person lying on
it will attempt to connect their internals to the attached tank.

Also since I had to adjust the table sprite, its lights now change color
based on whenever there's an active patient and an active operating
computer.

##### This is a commission for Improvedname/Ezel

## Why It's Good For The Game

Anesthetics are incredibly clunky to use, as either the person
themselves, or you as a surgeon, need to equip the tank onto whoever is
being operated on, then equip and activate the mask. This should make
them a bit less tedious, so maybe they'll actually see some use.

## Changelog
:cl:
add: You can now attach anesthetic tanks and breathing masks to surgical
tables.
/:cl: