Removes spacemove crashes/impacts at high speed (#92284)
## About The Pull Request

Something changed in spacemove loop code recently which caused infinite
recursions through throw code. This has been a rather bad mechanic
already, and I believe that it warrants removal alongside newtonian
zero-g physics in general - but for now lets just get rid of the part
that crashes the server.

## Why It's Good For The Game

We've pushed the speed limit to the point where you need to be firing a
SAW without a jetpack for a full minute to actually crash into
something, its a bad mechanic that doesn't bring much to the round
as-is.