The effects of flashbangs become less severe the further away you are from the flashbang when it detonates. (#92294)
## About The Pull Request

<img width="691" height="207" alt="flashbang agony zone"
src="https://github.com/user-attachments/assets/a4cc5fdb-19dc-4849-8c4a-2db1ac0fe212"
/>

In this handy chart, you can now see the impact of the flashbang based
on location.

**Dead Center:** While our Powergaming RD for Testing Pruposes is immune
to this, being on this tile when the flashbang goes off still hard stuns
you for a really long time even if you have ear protection. The bang
doesn't do anything in a vacuum.

**Pointblank:** This works as it does live, going through protections.
Though the bang doesn't work in a vacuum.

**Sweetspot:** This works as it does live, stunning and knocking you
down, but not if you're protected from the flash and the bang. This is
currently set to 3 tiles.

**Stagger Zone:** If you're vulnerable to the bang, you are staggered
and drop your items, rather than stunned and knocked down. If you're
vulnerable to the flash, you are dizzied and drop your items, rather
than stunned and knocked down.

**Safe Zone:** Basically this is outside of the flashbang's effective
range. The effective range is the same as live (7 tiles).

## Why It's Good For The Game

Flashbangs are a pretty easy room wide attack that acts as a one of our
more extreme gear checks. You either are entirely safe from this, or you
probably are going to eat shit to something just tossed into your line
of sight. Either you need to be able to disappear from the room
immediately, or the flashbang gets you.

Without nullifying the room attacking effect too severely, as well as
the mass disarming capabilities of the flashbang, this at least helps
provide some ways to counteract flashbangs without absolutely needing to
have the protections. You can play around the stagger and dizziness more
easily than you can the stuns.

The vacuum stuff is mostly because being hit by a flashbang's bang in a
vacuum is a bit absurd. The flash still works, however.

## Changelog
:cl:
balance: The effects of flashbangs become less severe the further away
you are from the flashbang.
balance: The bang effect of flashbangs cannot affect you in a vacuum.
The flash still functions in a vacuum.
/:cl:

---------

Co-authored-by: Aki Ito <11748095+ExcessiveUseOfCobblestone@users.noreply.github.com>